requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.22
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
pandas>=2.0.0
openpyxl>=3.1.0
selenium>=4.10.0
beautifulsoup4>=4.12.0
pdfplumber>=0.9.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0